+++ src/main/java/exchange/core2/cluster/model/binary/BinaryDataCommand.java
@@ -4,6 +4,6 @@ import exchange.core2.cluster.utils.BufferWritable;
 
 public interface BinaryDataCommand extends BufferWritable {
 
-    int getBinaryCommandTypeCode();
+    short getBinaryCommandTypeCode();
 
 }
